+++ HISTORY.rst
@@ -28,6 +28,7 @@ Unreleased
 * ALCN-1035: Open supplier promotion in form with a button from editable list view
 * ALCN-1035: Use purchase base price on supplier info onchange instead of list price
 * ALCN-1035: Allow to import supplier info with CNK code to link product
+* ALCN-1025: Add route for onorder to fridge, add routes for products, add routes per zone for nouveautés and achetés-vendus
 
 **Bugfixes**